add argparse

+++ lantmateriet/fetch.py
@@ -2,8 +2,7 @@ import os
 import requests
 import json
 import base64
-
-from pprint import pprint
+import argparse
 
 URL_BILD = "https://api.lantmateriet.se/stac-bild/v1"
 URL_HOJD = "https://api.lantmateriet.se/stac-hojd/v1"
@@ -51,7 +50,7 @@ def make_request(url, params: dict=None, headers: dict=None, type: str="json"):
         return response.content
 
 
-def download_and_save_tif(url, filepath, headers: dict|None=None):
+def download_and_save_tif(url, filepath, headers: dict|None=None, force_dl=False):
     try:
         if headers and "username" in headers.keys() and "password" in headers.keys():
             credentials = f"{headers["username"]}:{headers["password"]}".encode("utf-8")
@@ -62,15 +61,17 @@ def download_and_save_tif(url, filepath, headers: dict|None=None):
             headers.pop("username")
             headers.pop("password")
 
-        response = requests.get(url, headers=headers, stream=True)
-        response.raise_for_status()
 
-        os.makedirs(os.path.dirname(filepath), exist_ok=True)
-        with open(filepath, 'wb') as f:
-            for chunk in response.iter_content(chunk_size=8192):
-                f.write(chunk)
+        if force_dl or not os.path.exists(filepath):
+            response = requests.get(url, headers=headers, stream=True)
+            response.raise_for_status()
 
-        print(f"TIFF file downloaded and saved successfully to {filepath}")
+            os.makedirs(os.path.dirname(filepath), exist_ok=True)
+            with open(filepath, 'wb') as f:
+                for chunk in response.iter_content(chunk_size=8192):
+                    f.write(chunk)
+
+            print(f"TIFF file downloaded and saved successfully to {filepath}")
 
     except requests.exceptions.RequestException as e:
         print(f"Error downloading TIFF file: {e}")
@@ -91,18 +92,18 @@ def check_if_next_link_exists(search_data):
     return next_exists, out_links
 
 
-def main():
-    out_dir = "/home/alexandre/Desktop/download_lantmateriet"
+def main(search_url, out_dir, bbox):
     params = {
-        "bbox": "15.122,65.7567,15.3297,65.9744",
+        "bbox": bbox,
         "limit": 50
     }
     links = []
-    next_links = [URL_BILD+"/search"]
+    next_links = [search_url]
 
     while next_links:
         next = next_links.pop(0)
         search_data = make_request(next, params=params, type="json")
+
         for feat in search_data["features"]:
             links.append(feat["assets"]["data"]["href"])
 
@@ -111,10 +112,24 @@ def main():
             params = None
             next_links.extend(new_nexts)
 
-
     for link in links:
         output_file_path = os.path.join(out_dir, link.split("data/")[1])
         download_and_save_tif(link, output_file_path, headers=creds)
 
 if __name__ == "__main__":
-    main()
+    parser = argparse.ArgumentParser()
+    parser.add_argument("-o", help="Out directory")
+    parser.add_argument("--bbox", help="Bounding box in LONG,LAT,LONG,LAT coord.")
+    parser.add_argument("--type", help="bild / höjd", required=True)
+
+    args = parser.parse_args()
+
+    if args.type == "bild":
+        search_url = URL_BILD+"/search"
+    elif args.type == "höjd":
+        search_url = URL_HOJD+"/search"
+    else:
+        print(f"Bad type. Expecting bild or höjd, got {args.type}")
+        exit()
+
+    main(search_url, args.o, args.bbox)
